kununua_backend/requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
aniso8601==9.0.1
anyio==3.6.2
appdirs==1.4.4
asgiref==3.5.2
async-generator==1.10
async-timeout==4.0.2
asyncer==0.0.2
attrs==22.1.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
coloredlogs==15.0.1
Django==4.1.2
django-cors-headers==3.13.0
django-extensions==3.2.1
django-graphql-jwt==0.3.4
django-livereload-server==0.4
djangorestframework==3.14.0
exceptiongroup==1.0.4
fastapi==0.87.0
filetype==1.2.0
flatbuffers==23.1.21
frozenlist==1.3.3
graphene==3.1.1
graphene-django==3.0.0
graphql-core==3.2.3
graphql-relay==3.2.0
h11==0.14.0
httplib2==0.21.0
humanfriendly==10.0
idna==3.4
ImageHash==4.3.1
imageio==2.25.0
importlib-metadata==6.0.0
livereload==2.6.3
llvmlite==0.39.1
lxml==4.9.1
mpmath==1.2.1
multidict==6.0.4
networkx==3.0
numba==0.56.4
numpy==1.23.5
onnxruntime==1.13.1
opencv-python-headless==4.6.0.66
outcome==1.2.0
packaging==23.0
pandas==1.5.3
Pillow==9.3.0
pooch==1.6.0
promise==2.3
protobuf==4.21.12
psycopg2-binary==2.9.5
pydantic==1.10.4
PyJWT==2.6.0
PyMatting==1.1.8
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
python-multipart==0.0.5
pytz==2022.5
PyWavelets==1.4.1
rembg==2.0.30
requests==2.28.2
scikit-image==0.19.3
scipy==1.9.3
selenium==4.7.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
starlette==0.21.0
sympy==1.11.1
text-unidecode==1.3
tifffile==2023.1.23.1
tornado==6.2
tqdm==4.64.1
trio==0.22.0
trio-websocket==0.9.2
typing_extensions==4.4.0
urllib3==1.26.13
uvicorn==0.20.0
watchdog==2.1.9
wsproto==1.2.0
yarl==1.8.2
zipp==3.11.0
